+++ cmd/constants.go
@@ -22,7 +22,7 @@ var (
 )
 
 const (
-	version               = "v0.2.0"
+	version               = "v0.2.1"
 	deploymentSuccessMark = "----------deployment-success----------"
 	deploymentErrorMark   = "----------deployment-error----------"
 )