+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [NEXT_RELEASE]
+## [NEXT_RELEASE] ## [0.2.1] - 2017-03-14
 ### Added
 - read file `.teresaignore` to ignore some files in deploy process